Task Complete

+++ database/migrations/2022_05_13_220658_create_affiliates_table.php
@@ -18,7 +18,8 @@ return new class extends Migration
             $table->foreignId('user_id');
             $table->foreignId('merchant_id');
             // TODO: Replace me with a brief explanation of why floats aren't the correct data type, and replace with the correct data type.
-            $table->float('commission_rate');
+            //decimal datatype help to preserve exact precision, while the float dosnt.
+            $table->decimal('commission_rate',8, 2);
             $table->string('discount_code');
             $table->timestamps();
         });
